feat(tabs): Rewrite of PapillonTabNavigator for performance
