+++ src/router/navigator/atoms/TabItem.tsx
@@ -1,0 +1,129 @@
+import * as React from "react";
+import { useCurrentAccount } from "@/stores/account";
+import { useNavigationBuilder, useTheme } from "@react-navigation/native";
+import { StyleSheet, View, Text, Platform } from "react-native";
+import { useSafeAreaInsets } from "react-native-safe-area-context";
+import LottieView from "lottie-react-native";
+import colorsList from "@/utils/data/colors.json";
+import { Pressable } from "react-native-gesture-handler";
+import * as Haptics from "expo-haptics";
+import Reanimated, { FadeIn, FadeOut, LinearTransition, ZoomIn } from "react-native-reanimated";
+import { anim2Papillon } from "@/utils/ui/animations";
+
+const TabItem: React.FC<{
+  route: any;
+  descriptor: any;
+  navigation: any;
+  isFocused: boolean;
+}> = ({ route, descriptor, navigation, isFocused }) => {
+  const theme = useTheme();
+
+  const { options } = descriptor;
+  const label = options.tabBarLabel !== undefined ? options.tabBarLabel : options.title !== undefined ? options.title : route.name;
+
+  const onPress = () => {
+    const event = navigation.emit({
+      type: "tabPress",
+      target: route.key,
+      canPreventDefault: true
+    });
+
+    if (!isFocused && !event.defaultPrevented) {
+      navigation.navigate(route.name);
+    }
+
+    lottieRef.current?.play();
+    Haptics.impactAsync(Haptics.ImpactFeedbackStyle.Light);
+
+  };
+
+  const onLongPress = () => {
+    navigation.emit({
+      type: "tabLongPress",
+      target: route.key
+    });
+  };
+
+  const lottieRef = React.useRef<LottieView>(null);
+
+  const autoColor = colorsList.filter(c => c.hex.primary === theme.colors.primary)[0];
+
+  const tabColor = isFocused ?
+    (theme.dark ? autoColor.hex.lighter : autoColor.hex.dark) : (theme.dark ? "#656c72" : "#8C9398");
+
+  return (
+    <Reanimated.View
+      key={"tab-tabButton-" + route.key}
+      style={[styles.tabItemContainer]}
+      layout={anim2Papillon(LinearTransition)}
+    >
+      <Pressable
+        accessibilityRole="button"
+        accessibilityState={isFocused ? { selected: true } : {}}
+        accessibilityLabel={options.tabBarAccessibilityLabel}
+        testID={options.tabBarTestID}
+        onPress={onPress}
+        onLongPress={onLongPress}
+        style={[styles.tabItem]}
+      >
+        <Reanimated.View
+          entering={anim2Papillon(ZoomIn)}
+          exiting={anim2Papillon(FadeOut)}
+        >
+          {options.tabBarLottie && (
+            <LottieView
+              loop={false}
+              source={options.tabBarLottie}
+              colorFilters={[{
+                keypath: "*",
+                color: tabColor,
+              }]}
+              style={[
+                {
+                  width: 26,
+                  height: 26,
+                }
+              ]}
+              ref={lottieRef}
+            />
+          )}
+        </Reanimated.View>
+
+        <Reanimated.Text
+          style={[
+            styles.tabText,
+            { color: tabColor },
+            Platform.OS === "android" && { fontFamily: undefined }
+          ]}
+          numberOfLines={1}
+          entering={anim2Papillon(FadeIn)}
+          exiting={anim2Papillon(FadeOut)}
+        >
+          {label}
+        </Reanimated.Text>
+      </Pressable>
+    </Reanimated.View>
+  );
+};
+
+const styles = StyleSheet.create({
+  tabItemContainer: {
+    flex: 1,
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  tabItem: {
+    flexDirection: "column",
+    alignItems: "center",
+    justifyContent: "center",
+    padding: 0,
+    gap: 4,
+  },
+  tabText: {
+    fontSize: 13,
+    textAlign: "center",
+    fontFamily: "medium",
+  },
+});
+
+export default React.memo(TabItem);
