+++ src/router/navigator/tabs.tsx
@@ -1,0 +1,59 @@
+import React, { useMemo } from "react";
+import { useCurrentAccount } from "@/stores/account";
+import { useNavigationBuilder, useTheme } from "@react-navigation/native";
+import { StyleSheet, View, Platform } from "react-native";
+import { useSafeAreaInsets } from "react-native-safe-area-context";
+import TabItem from "./atoms/TabItem";
+
+import Reanimated, { LinearTransition } from "react-native-reanimated";
+
+const PapillonNavigatorTabs: React.FC<Omit<ReturnType<typeof useNavigationBuilder>, "NavigationContent">> = ({ state, descriptors, navigation }) => {
+  const theme = useTheme();
+  const insets = useSafeAreaInsets();
+  const account = useCurrentAccount((store) => store.account);
+
+  const allTabs = useMemo(() => state.routes, [state.routes]);
+  const tabs = account?.personalization.tabs?.filter((tab) => tab.enabled)
+    ?.map((tab) => allTabs.find((route) => route.name === tab.name))
+    .filter(Boolean) || allTabs.slice(0, 5);
+
+  return (
+    <Reanimated.View
+      style={[
+        styles.tabBar,
+        {
+          backgroundColor: theme.colors.card,
+          borderColor: theme.colors.border,
+          paddingBottom: insets.bottom + 10,
+        },
+        Platform.OS === "android" ? styles.tabBarAndroid : styles.tabBarIOS,
+      ]}
+    >
+      {tabs.map((route, index) => (
+        <TabItem
+          key={route.key}
+          route={route}
+          descriptor={descriptors[route.key]}
+          navigation={navigation}
+          isFocused={allTabs.indexOf(route) === state.index}
+        />
+      ))}
+    </Reanimated.View>
+  );
+};
+
+const styles = StyleSheet.create({
+  tabBar: {
+    flexDirection: "row",
+    paddingHorizontal: 8,
+    paddingTop: 8,
+    zIndex: 1000,
+    borderTopWidth: 0.5,
+  },
+  tabBarAndroid: {
+    elevation: 10,
+  },
+  tabBarIOS: {},
+});
+
+export default PapillonNavigatorTabs;
